replace Test Writer to separate project for tests extensions and create project for kitchen service tests
