+++ MessageDriven.sln
@@ -30,6 +30,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{ADE02B1B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Restaurant.Booking.Tests", "Restaurant\Restaurant.Booking.Tests\Restaurant.Booking.Tests.csproj", "{EC595166-2B0B-4A76-9D70-7256779298D9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Restaurant.Kitchen.Tests", "Restaurant\Restaurant.Kitchen.Tests\Restaurant.Kitchen.Tests.csproj", "{58DFE90F-E37E-4C2D-A5F4-C7B04365B051}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestsExtensions", "Restaurant\TestsExtensions\TestsExtensions.csproj", "{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -84,6 +88,14 @@ Global
 		{EC595166-2B0B-4A76-9D70-7256779298D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC595166-2B0B-4A76-9D70-7256779298D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC595166-2B0B-4A76-9D70-7256779298D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{58DFE90F-E37E-4C2D-A5F4-C7B04365B051}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{58DFE90F-E37E-4C2D-A5F4-C7B04365B051}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{58DFE90F-E37E-4C2D-A5F4-C7B04365B051}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{58DFE90F-E37E-4C2D-A5F4-C7B04365B051}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{61297776-527A-4F81-90A6-397996EE3F75} = {E0BA00B7-42BC-49BC-BB50-B058D3DDCD3C}
@@ -95,5 +107,7 @@ Global
 		{674364E1-6F40-4F62-941B-917ABB4C2447} = {DB5AB717-2339-4AFB-8B31-5E6778F6AA56}
 		{ADE02B1B-4B31-489E-8731-F1600BB491AA} = {DB5AB717-2339-4AFB-8B31-5E6778F6AA56}
 		{EC595166-2B0B-4A76-9D70-7256779298D9} = {ADE02B1B-4B31-489E-8731-F1600BB491AA}
+		{58DFE90F-E37E-4C2D-A5F4-C7B04365B051} = {ADE02B1B-4B31-489E-8731-F1600BB491AA}
+		{E19A2D3C-256C-469B-AE81-FA2ABA2EFC39} = {ADE02B1B-4B31-489E-8731-F1600BB491AA}
 	EndGlobalSection
 EndGlobal
