+++ Restaurant/Restaurant.Kitchen.Tests/Restaurant.Kitchen.Tests.csproj
@@ -9,7 +9,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="MassTransit.Quartz" Version="8.0.6" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
@@ -23,9 +22,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Restaurant.Booking\Restaurant.Booking.csproj" />
-      <ProjectReference Include="..\Restaurant.Kitchen\Restaurant.Kitchen.csproj" />
-      <ProjectReference Include="..\TestsExtensions\TestsExtensions.csproj" />
+        <ProjectReference Include="..\Restaurant.Kitchen\Restaurant.Kitchen.csproj" />
+        <ProjectReference Include="..\TestsExtensions\TestsExtensions.csproj" />
     </ItemGroup>
-
 </Project>
