+++ Restaurant/Restaurant.Kitchen.Tests/Usings.cs
@@ -2,7 +2,7 @@ global using Xunit;
 global using MassTransit;
 global using MassTransit.Testing;
 global using Microsoft.Extensions.DependencyInjection;
-global using Restaurant.Booking.Consumers;
 global using Restaurant.Messages;
 global using Xunit.Abstractions;
 global using TestsExtensions;
+global using Restaurant.Kitchen.Consumers;
